supply-chain-planning-system/requirements.txt
# Supply Chain Planning System - Master Orchestrator
# Python 3.11+

# Core Dependencies
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# Configuration
pyyaml>=6.0
pydantic>=2.0.0

# Machine Learning (from sub-modules)
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=4.0.0
statsmodels>=0.14.0

# Optimization (from sub-modules)
pulp>=2.7.0
ortools>=9.6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Dashboard
streamlit>=1.28.0

# Task Scheduling (for orchestration)
schedule>=1.2.0
celery>=5.3.0
redis>=5.0.0

# Data Validation
pandera>=0.17.0

# Logging & Monitoring
structlog>=23.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Code Quality
black>=23.7.0
flake8>=6.0.0
mypy>=1.4.0

# Type Stubs
pandas-stubs>=2.0.0
types-PyYAML>=6.0.0
